Modify fuse tests
